+++ app/src/main/java/prac/tanken/shigure/ui/subaci/ui/playlist/model/PlaylistUpsertIntent.kt
@@ -2,5 +2,5 @@ package prac.tanken.shigure.ui.subaci.ui.playlist.model
 
 sealed class PlaylistUpsertIntent {
     data object Insert : PlaylistUpsertIntent()
-    data class Update(val originalId: Int) : PlaylistUpsertIntent()
+    data class Update(val originalId: Long) : PlaylistUpsertIntent()
 }
